docs: TokenKit
